refactor RoomComponent
